Refactored and cleaned up command.rs and subscriber.rs
