+++ src/builders.rs
@@ -1,7 +1,7 @@
 use std::sync::{Arc, Mutex};
 use crate::helpers::{Deferrable, Deferred};
-use crate::io::{Comparison, Device, GenericInput, IdType, InputType, IOKind, IOType, BaseCommandFactory, SimpleNotifier, ThresholdNotifier};
-use crate::io::{Publisher, PublisherInstance};
+use crate::io::{BaseCommandFactory, Device, GenericInput, IdType, InputType, IOKind, IOType};
+use crate::io::{Comparison, ThresholdNotifier, SimpleNotifier, Publisher, PublisherInstance};
 use crate::settings::Settings;
 use crate::storage::OwnedLog;
