+++ src/io/action.rs
@@ -2,8 +2,12 @@ mod command;
 mod subscriber;
 mod publisher;
 mod types;
+mod subscribers;
+mod commands;
 
 pub use command::*;
+pub use commands::*;
 pub use subscriber::*;
+pub use subscribers::*;
 pub use publisher::*;
 pub use types::*;
