+++ src/io/action/publisher.rs
@@ -1,8 +1,16 @@
+//! Implement observer design pattern to implement control system based off of polling of `Input` objects
+//!
+//! # Operation
+//! `Publisher` objects should be stored a struct which implements `Input`. When `Input::read()` is called,
+//! `Input::publisher().notify()` should also be called as well. `notify()` should thereby call
+//! `Subscriber::evaluate()` on any listeners.
+//!
+//! The goal of a dedicated `Publisher` implementation being stored as a field is to add a layer of indirection
+//! between `Input` and `Subscriber` to serve as a mediator.
+
 use std::sync::{Arc, Mutex};
 use crate::helpers::{Deferrable, Deferred};
 use crate::io::{IOEvent, SubscriberType};
-
-/// Implement observer design pattern to implement control system based off of polling of `Input` objects
 
 pub trait NamedRoutine {
     fn name(&self) -> String;
@@ -16,6 +24,7 @@ pub trait Publisher: Deferrable {
     fn notify(&mut self, data: &IOEvent);
 }
 
+/// Concrete instance of publisher object
 #[derive(Default, Clone)]
 pub struct PublisherInstance {
     subscribers: Vec<Deferred<SubscriberType>>
@@ -28,6 +37,7 @@ impl Publisher for PublisherInstance {
         self.subscribers.push(subscriber)
     }
 
+    /// Call `Subscriber::evaluate()` on all associated `Subscriber` implementations.
     fn notify(&mut self, data: &IOEvent) {
         for subscriber in self.subscribers.iter_mut() {
             // TODO: `IOEvent` shall be sent to `OutputDevice` and shall be logged
